chore: Update test code
To fix test order

Related to: #40
