+++ ctrlers/gov/2_voting_test.go
@@ -120,6 +120,7 @@ func TestVoting(t *testing.T) {
 func TestMajority(t *testing.T) {
 	prop, xerr := govCtrler.ReadProposal(trxCtxProposal.TxHash)
 	require.NoError(t, xerr)
+	require.NotNil(t, prop)
 
 	opt := prop.UpdateMajorOption()
 	require.Nil(t, opt)
@@ -134,11 +135,12 @@ func TestMajority(t *testing.T) {
 
 		prop, xerr := govCtrler.ReadProposal(trxCtxProposal.TxHash)
 		require.NoError(t, xerr)
+		require.NotNil(t, prop)
 
 		votedPowers += stakeHelper.PowerOf(c.txctx.Tx.From)
 		if votedPowers >= prop.MajorityPower {
 			opt := prop.UpdateMajorOption()
-			require.NotNil(t, opt)
+			require.NotNil(t, opt, votedPowers, prop.MajorityPower)
 			require.EqualValues(t, prop.MajorOption, opt)
 			require.Equal(t, votedPowers, opt.Votes())
 		} else {
@@ -159,6 +161,7 @@ func TestMajority(t *testing.T) {
 
 		prop, xerr := govCtrler.ReadProposal(trxCtxProposal.TxHash)
 		require.NoError(t, xerr)
+		require.NotNil(t, prop)
 
 		opt := prop.UpdateMajorOption()
 		require.NotNil(t, opt)
